DE345531 - Disallow using Restman to delete root folder (#822)
* DE345531 - Don't allow deleting root folder

* DE345531 - Remove lambda so build will pass

* DE345531 - Remove unneeded throws declarations

* DE345531 - PR fixes

+++ modules/gateway/assertions/GatewayManagementAssertion/src/main/java/com/l7tech/external/assertions/gatewaymanagement/server/rest/factories/impl/FolderAPIResourceFactory.java
@@ -73,9 +73,12 @@ public class FolderAPIResourceFactory extends WsmanBaseResourceFactory<FolderMO,
     }
 
     public void deleteResource(@NotNull String id, boolean forceDelete) throws ResourceFactory.ResourceNotFoundException {
-        if(forceDelete){
+        if (forceDelete) {
+            if (Folder.ROOT_FOLDER_ID.equals(Goid.parseGoid(id))) {
+                throw new ResourceFactory.ResourceAccessException(new ConstraintViolationException("Cannot delete root folder"));
+            }
             deleteRecursively(id);
-        }else {
+        } else {
             super.deleteResource(id);
         }
     }
